+++ app/src/Model/Oauth2Providers.php
@@ -1,0 +1,21 @@
+<?php
+namespace Dappur\Model;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Oauth2Providers extends Model {
+
+    protected $table = 'oauth2_providers';
+    protected $primaryKey = 'id';
+    protected $fillable = [
+    	'name',
+    	'slug', 
+    	'scopes',
+        'authorize_url', 
+    	'token_url',
+        'resource_url',
+    	'button',
+        'login',
+        'status'
+    ];
+}