+++ phinx.php
@@ -1,9 +1,12 @@
 <?php
 require __DIR__ . '/vendor/autoload.php';
+
 $settings = file_get_contents( __DIR__ . '/settings.json');
 $settings = json_decode($settings, TRUE);
 
 $dbconf = $settings['db'][$settings['environment']];
+
+
 
 return [
   	'paths' => [